style<CategoryDto> 디렉터리 구조 이동
