--- src/main/java/com/gulbi/Backend/domain/rental/product/dto/ProductRegisterRequestDto.java
@@ -1,24 +1,0 @@
-package com.gulbi.Backend.domain.rental.product.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class ProductRegisterRequestDto {
-    private String tag;
-    private String title;
-    private String productName;
-    private String price;
-    private String sido;
-    private String sigungu;
-    private String bname;
-    private String description;
-    private String bcategoryId;
-    private String mcategoryId;
-    private String scategoryId;
-}
